raw version of the logic
ready to use. The logic is save in one file. We have 2 contracts to deploy.

Datas can be pass as tuple. No duplicat. The main contract will throw an error if it is the case. 

Contracts for articles can be generated based on the input. These new contracts can be retrieved. 


request with the url only can also be evaluate. As response the blockchain give back all contracts associate to the url.

Security and exceptions have to be done.